Conflicts:
	Projects/iPhoneFramework/Framework.xcodeproj/project.pbxproj

Former-commit-id: dee9240e156823ff720340d89576422193791acd

+++ Sources/Internal/FileSystem/YamlEmitter.cpp
@@ -95,7 +95,7 @@ int write_handler(void *ext, unsigned char *buffer, size_t size)//yaml_write_han
 {
     File *yamlFile = static_cast<File *>(ext);
 
-    uint32 bytesWritten = yamlFile->Write(buffer, size);
+    uint32 bytesWritten = yamlFile->Write(buffer, static_cast<uint32>(size));
 
     return (size == bytesWritten) ? 1 : 0;
 }
